new: Added ReDoc customizations for Linode
This included a redoc standalone build from [this
fork](https://github.com/dnknapp/ReDoc#redoc-options-object) that has
customizations done specifically for Linode.

In this PR I've committed the standalone js file directly into this
repo.  As such, we're losing any benefits of using the file behind the
CDN, and are assuming responsibility for maintaining our fork of ReDoc
(i.e. if upstream updates, we'll have to pull the changes in and apply
our patches if we want them).  I decided against making a builder for
ReDoc since building _one file_ is relatively trivial, and I don't want
to make it more complicated than necessary.