Added automatic versioning through git tags and Jenkins
Versions are taken _exclusively_ from git tags, unlike other repos where
we include a patch version based on commits since the last tag.  This is
because we will probably want to tightly control incrementing patch
versions of the spec, and if we made four commits of typo-fixes we
probably don't want to bump the version by 0.0.4.  As such, packages are
going to be versioned the same as the spec, but with a
`-${commits_since_tag}` appended to the tag's version number.  All tags
going forward should match the format of the tag I am applying to this commit:
`4.0.0`.