ref: Made all document links absolute
ARB-695 #done

The benefit of having all urls absolute is that the spec will continue
to be accurate even if it is not presented at developers.linode.com.
Also updated the Jenkinsfile to sub the URLs out depending on the
environment we're building it for.

**NOTE** This will break links until the nginx config in dev is updated
to move the page to its correct eventual address