Clarifies requirements for /profile.put
@ctarquini forwarded a report from HackerOne claiming that arbitrary
third party application could update profile - and this is true, but
expected.  There is ongoing discussion about making a more specific
OAuth Scope for profile, but the existing language in the docs made the
reporter think that third party apps should _never_ be able to access
the /profile.put operation, where indeed they can with the right scope.