Update POST /account/oath-clients to get public parameter to show
The `public` parameter was already in the yaml, but was not showing up. This is because it is given `readOnly: true` in the corresponding `components/schemas` section. Will Smith pointed out that it would show up if it was given `readOnly: false`, and suggested copying the block both above and below the section's $ref. This is valid schema, and does not break reDoc or the CLI.