Add /account/cancel endpoint (#306)
* Add /account/cancel endpoint

This endpoint cancels an account. The only request body parameter is the `comments` field, which I've suggested is a place for the user to return the reason for the cancellation / any other feedback.

* Fixed x-linode-cli-command

* Added statement to clarify that cancel action

Added a statement to ensure users are aware that this action will cause Linode to charge the credit card on file for the remaining balance.

* Updated endpoint and comment param descriptions.

Tried to make it more clear that Linode will not be able to proceed with the account cancellation if they are unable to charge a credit card.  Made the `comment` param a bit more open ended while still encouraging feedback by changing `the reason` to `any reason`.

* Changed exit survey link to dummy link

This is to prevent a customer from sending bogus survey results.