new: Added documentation for the object-storage/buckets endpoints
This is a draft; please review carefully and do not merge without
populating placeholder links.

These endpoints have existed for some time, but are finally being
documented (the last two are shipping on Monday).  I _do not_ advise
actual users to use these endpoints; they're mostly for the benefit for
cloud.  In keeping with API philosophy, they _are_ available to all
customers, but the s3 API will almost certainly be more useful in all
cases, which is why using these endpoints is discouraged.