Added new CLI-specific parameters to backups endpoint
This goes with https://github.com/linode/linode-cli/pull/156

The CLI had a hard time parsing the output of this endpoint well, so I
made the above CLI PR to use these two new spec extensions to give it a
hint as to how to parse these responses.  They shouldn't change anything
in the docs at all.