Allow CLI to accept k8s version when creating clusters
This is for https://github.com/linode/linode-cli/pull/177

Right now, the "version" argument accepted when creating an LKE Cluster
translates in the CLI to `--version`, which the CLI interprets to mean
it should print its own version and exit.

For this reason, a new spec extension has been added,
`x-linode-cli-name`, which is allowed in schemas to define a new name
for the CLI argument for any POST/PUT attribute.  To keep things
consistent between the API and CLI, this should be used only in extreme
circumstances (such as noted above).