Filled in IPv6Range object
Closes https://github.com/linode/linode-cli/issues/207

It was reported above that the CLI wasn't showing complete output for
IPv6 Ranges compared to IPv6 Pools.  Turns out this is because the docs
were missing several fields for ranges.

This PR adds the missing fields to IPv6Ranges.  I based the descriptions
and values off of the equivalent fields in IPv6Pool.