Changed language for app expiry
Changed the description to be more clear as to what steps to
take based on what the return value of expiry is. If it's a public app
no refresh token is created. If it's a private app, a refresh token is
created and never expires, so the API is correct in its map output.