new: Documented POST /images/upload endpoint
Needed for https://github.com/linode/linode-cli/pull/234

This endpoint is currently in beta, and is not yet available for all
customers.

The CLI needs this endpoint to be documented for the convenience plugin
to build and work correctly, and we do not expect the spec of this
endpoint to change in the future.  I included information about how to
correctly use the endpoint, and how to enroll in the beta, but I am open
to any changes or suggestions.