Add contentEncoding and contentMediaType for Kubeconfig
This is related to https://github.com/linode/linode-cli/issues/247

The kubeconfig returned from `GET /lke/clusters/{clusterId}/kubeconfig`
is a base64-encoded YAML config.  The OpenAPI standard now supports this
[with the standard JSON Schema fields](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#considerations-for-file-uploads).

This change notes the encoding of the `kubeconfig` field, which the CLI
may use to decode the base64 data and display the actual YAML for
convenience (see linked CLI PR below).