Fix linode-cli firewalls create
This closes https://github.com/linode/linode-cli/issues/249

It looks like the CLI wanted every single element of a firewall to come
in as a separate flag, but inbound/outbound rules include lists of
objects, and the CLI doesn't really support those as flags on the
command line.  The docs suggested the entire rules object could be
provided as a JSON string, but I decided to meet in the middle and have
`--rules.outbound` and `--rules.inbound` accept JSON strings, and have
everything else be a separate flag.  Details are in the issue linked
above.