bug: Fixed CLI example for `firewalls rules-update` action
Closes https://github.com/linode/linode-cli/issues/254

I was taking a pass at open CLI issues and found that while this one was
fixed upstream, the example it referenced in the docs was still not
working.  With a little tinkering, I believe this fixes it.

I'm not sure if the values I put in here actually make sense for a real
use-case, I just picked valid values.  Please modify them if you think
they're silly or could be more meaningful.